+++ src/server/createController.ts
@@ -1,10 +1,14 @@
-import { HistoryLocation, Context } from '../share/type'
-import { ServerController, ServerControllerConstructor } from './type'
+import type {
+  ServerController,
+  ServerControllerConstructor,
+  HistoryLocation,
+  Context
+} from "./index";
 
 export default function createController(
   c: ServerControllerConstructor,
   location: HistoryLocation,
   context: Context
 ): ServerController {
-  return new c(location, context)
+  return new c(location, context);
 }
